Ensure Provider is populated when loading an existing workload cluster
